commented git domain models

+++ springboot/src/test/java/ca/jrvs/apps/trading/dao/AccountDaoIntTest.java
@@ -46,7 +46,7 @@ public class AccountDaoIntTest {
     traderDao.save(savedTrader);
     savedAccount = new Account();
     savedAccount.setTraderId(1);
-    savedAccount.setAmount(0);
+    savedAccount.setAmount(0d);
     accountDao.save(savedAccount);
   }
 
@@ -62,7 +62,7 @@ public class AccountDaoIntTest {
         .newArrayList(accountDao.findAllById(Collections.singletonList(savedAccount.getId())));
 
     assertEquals(1, accountList.size());
-    assertEquals(savedAccount.getAmount().doubleValue(), accountList.get(0).getAmount());
+    assertEquals(savedAccount.getAmount(), accountList.get(0).getAmount());
     assertEquals(savedAccount.getTraderId(), accountList.get(0).getTraderId());
   }
 }
